Move components common to all create pages to a dedicated folder
